Refactor suggested music components for modularity
Extract SongSearchInput and SongList into separate components to improve maintainability and reuse. Centralize Spotify API logic in a new SpotifyService for cleaner data handling. Simplify CWMSuggestedMusic by integrating the new components.
